+++ src/card/sets/y2013/mod.rs
@@ -1,5 +1,6 @@
 //! Card sets first released in 2013.
 
+pub(super) mod commander_2013;
 pub(super) mod dragons_maze;
 pub(super) mod gatecrash;
 pub(super) mod magic_2014;